Cite the source of the movie dataset
Also remove unnecessary import in models.py and the test endpoint in
controllers.py.

No more work should be done on this branch. Modify in future branches.
Note that the database table for movies changed a bit in terms of field
names and the fact that the "source" field was removed.