Fix user account bricking if movie doesn't exist
If a user tries to add a movie with a title that the omdb api doesn't
recognize then the whole account gets bricked. The fix, for now, is to
acknowledge the request failure from omdb and just add it to the watch
list. Acknowledge that it doesn't exist in omdb on the movie listing.

This is kind of unappealing. Ideally the easiest thing would be to not
allow people to send the form / ask them to edit their form if the request
to the api fails.