Fix non existant movie issue in feed
Fix this issue by just checking if the response from omdbapi failed.
If so then ignore it and in the tooltip, mention that there was an
error.

Again, like with the watch list handling of this error, there should
be a mechanism in place that checks to see if the movie exists
in the omdbapi before allowing the form to be accepted.