Fix indentation and missing closing tags
Fix indentation in the feed and index tabs with 2 space indentation.
Also comment out the line about the movie entries being placed in
columns.

I wasn't sure what the feed layout was going to be because it looked
like columns were desired based on the code written. But the actual
effect was similar to the look of the watch list were each movie
entry was displayed as block. For the meantime I commented out that
columns line. But if you remove the comment then it will have the
column setup. Needs work though if you do want a column setup.