Show review comments when hovering over user icon
When the user is in their watch list and there are comments for
the review the user posted, if the user hovers over the commentors
icon then their comment is displayed.

I can't stress enough how inefficient this solution is. It requires
a bit of time to optimize. Probably wont have time though.