Level watchlist tab title and add button
Replace the bulma class "is-pulled-right" on the add movies button
in the watchlist tab to instead make use of the "level-item" class.
Surround the title and button with appropriate elements.

In the future, use level-left and level-right if you want to move
things around. is-pulled-right and is-pulled-left make use of floats.
Floats, from what I remember, get nasty really quick if you use too
many of them.