Add review text in movie entries in the feed tab
Review text should also show up in the feed tab similar to how it
shows up in the watchlist tab. The header for each movie entry is
either "My Review" or "x's Review"

I think that maybe we should only display other peoples movies.
Not our own. But for now I'm sticking with the original vision.