Restrict autoreload to previous commit description
In the previous commit, the intention was to only reload when
a user didn't have a post there previously. But what ended up
happening is that the post comment action will always reload.
The reload will only happen if the user is entering a comment
on a particular movie listing for the first time.

This is a really ugly, maybe cheap way of doing things. But I
didn't plan this far ahead.